+++ lib/src/main/java/com/formance/formance_sdk/models/shared/AssetHolder.java
@@ -6,16 +6,18 @@ package com.formance.formance_sdk.models.shared;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigInteger;
+
 public class AssetHolder {
     @JsonProperty("assets")
-    public java.util.Map<String, Long> assets;
+    public java.util.Map<String, BigInteger> assets;
 
-    public AssetHolder withAssets(java.util.Map<String, Long> assets) {
+    public AssetHolder withAssets(java.util.Map<String, BigInteger> assets) {
         this.assets = assets;
         return this;
     }
     
-    public AssetHolder(@JsonProperty("assets") java.util.Map<String, Long> assets) {
+    public AssetHolder(@JsonProperty("assets") java.util.Map<String, BigInteger> assets) {
         this.assets = assets;
   }
 }
